Made seperate kernel for block offset calculations nth element

+++ rocprim/include/rocprim/device/device_nth_element.hpp
@@ -56,8 +56,8 @@ ROCPRIM_INLINE hipError_t nth_element(void*          temporary_storage,
     Key*    tree    = nullptr;
     size_t* buckets = nullptr;
     // Maximum items per block 65535
-    unsigned short* buckets_per_block = nullptr;
-    size_t*         nth_element_data  = nullptr;
+    size_t* buckets_per_block = nullptr;
+    size_t* nth_element_data  = nullptr;
     // Maximum of 256 buckets
     unsigned char* oracles          = nullptr;
     bool*          equality_buckets = nullptr;
